+++ CHANGELOG.md
@@ -14,6 +14,8 @@ This changelog is following the recommended format by [keepachangelog](https://k
 
 ### Fixed
 
+- Issue when policy provisioning name is empty (cf. [#113](https://github.com/yannick-beot-sp/vscode-sailpoint-identitynow/issues/113)). If no name is provided, the usage type is used for the label
+
 ### Removed
 
 ## [1.3.15] - 2025-04-06